Clean up geometry module boundaries
